Give the configuration its own module
Knowing where the snapshots live, or how many minutes the scheduler
waits, meant importing plugin.py and with it the whole HTTP server.
api.py did it for a socket path, scheduler.py for two file names and a
delay, cli.py for six constants. Those settings now live in config.py,
which knows nothing about a volume or a request.

api.py keeps its import of plugin.py, written out this time, because that
import is also what posts the routes on the Bottle application. Nothing
else in the daemon imports plugin.py any more, so without that line
buttervolume run would serve an application without a single route, and
say nothing about it. The test suite cannot see that happen, since it
imports plugin.py on its own account, so a new test asks a fresh
interpreter the only question that matters: importing the API, are there
routes?
